Form dialogs for STAT and LIGH are finished
Form-editing dialogs for STAT and LIGH are now finished.

STAT: Scrapped the "Auto-populate" button; its behavior in the CK was rather silly anyway.

STAT: We now emit a save error if a LOD mesh path is longer than the allowed maximum. The GUI should now enforce that maximum length as well.

LIGH: Improved how we map implementation-detail flags to friendly enums, so it's harder to accidentally try to use those flags within the frontend, and so that we do the same mapping for the light flicker type.

LIGH: We now properly support the ICON subrecord (we were forgetting to load it) and the MICO subrecord (we hadn't decoded it at all).

Improvements to ui::bind: we now support groupboxes with checkboxes, and DKGameFilePickers.