UI-side renderer throttle; backend fixes; NOTE and FACT UI
Modified the NPC_ loader to coalesce duplicate faction membership entries, as the game also does this.

FACT: We now load faction icons; we'd forgotten this previously.

FACT: Fixed a crash when cloning forms, including for working copies when editing in the GUI.

Added form-editing dialogs for FACT and NOTE.

Modified the Render Window to throttle repainting to 60 FPS. We already had a throttle within the renderer itself, but I'd forgotten about us essentially spamming window repaints, and it was killing performance system-wide while the program was open.