Faction dialog seems just about done now
The faction dialog now performs initial updates for the enable state of the interfaction relationship and rank widgets. Simply put: when you initially open the dialog and have no other faction or rank selected, the widgets to edit your selection will now be disabled.

The faction dialog now has a sub-dialog for editing the vendor services location.

The faction dialog no longer has a button to preview the vendor services location, as some location values aren't easy to translate into a specific place to pop open in the Render Window.