critical fixes to selection, models
Wow, it's so cool and normal that QItemSelectionModel::setCurrentIndex will disobey the selection behavior and mode of the associated view, e.g. selecting a single cell in a table where selections should be whole rows!

Wow, it's so cool and normal that Qt offers no actual documentation or reference information for the proper use of QAbstractItemModel::changePersistentIndexList and friends! It's also super cool that the situation is so bad that 90% of code samples on the web that concern implementing QAbstractItemModel::sort are completely incorrect because they don't even update QPMIs at all!

Very cool!!! :D :D :D :D :D