Started work on inventory-editor (TESContainer) UI
Started work on a UI for editing inventories (e.g. CONT, NPC_, ReferenceAlias). The preview dialog is directly copied and pasted from the leveled list preview dialog with basically no changes.

DKNavmeshGenerationImportOptionPicker: tab focus fix. Need to forward this fix to the Qt Designer plug-in and then fix tab ordering in the dialogs that used the widget.

Created a Qt Designer file for Container forms. It's incomplete; need to finish the DKInventoryForm widget.