Form type icons
Not sure if I'll use these yet. They're also not terribly consistent (e.g. line thickness), and the sizing on some feels a bit jank.

Having icons in the listviews for leveled lists and inventories would be a nice touch, but I'm just not fully confident in these, and adding icons for every form type is low-prio right now anyway.