DKFormListPane extra column functionality; ActorBase work
Added an editor subsystem for game settings, to automate some of the boilerplate involved in getting a setting value (without editing DovahKitCore and forcing half the program to recompile).

Enhanced DKFormListPane: it is now possible to add extra columns to the widget, with these columns using caller-supplied names and getters. We cache the retrieved values and handle the logic of when to update them.

Identified the unknown morph float on ActorBase based on reverse-engineering: it's a magnitude for VampireMorph. If I'm properly understanding what I saw, this morph would be in the range [0, 1] rather than the usual [-1, 1].

Improvements to the ActorBase UI: we now generate the drop-downs for indexed face morphs; we now properly save faction changes; we now use a DKFormListPane for the package list; we now modify visibility of the FaceGen-related tabs when the race is changed (and on initial load) based on whether it's a FaceGen race.