Hardcoded form IDs; RACE error reporting; ActorBase UI work
The game setting subsystem is now initialized at startup, and now has a signal for when a setting value changes. The signal does not pass the new value as an argument. This is intentional: it is expected that the code you'll hook up to the signal would just re-query the value anyway (either that, or you'd have to write two versions of whatever you're doing: one for init and one for changes).

Moved all hardcoded form ID constants to a separate file from the hardcoded form factory.

Added code to properly stringify and report warnings and errors from RACE forms.

The ActorBase UI should now handle PC Level Mult more accurately.

The ActorBase UI has code to handle previewing actors' attribute values. It's 99% correct. We're a point off when testing Brelyna Maryon; we exactly follow the algorithm described on the CK wiki's Class article and that evidently is not correct.

Restored "What's This?" text for the ActorBase package list (lost in a previous commit), and added more detail. Added "What's This?" text for a few other widgets.

ActorBaseSkillsModel now uses the game setting subsystem.

Documented plans for redesigning localized string handling post-launch.