ActorBase work
ActorBase UI: improved "What's This?" text for widgets related to stats and creature sounds. The latter involved quite a bit of reverse-engineering (all just to figure out that the whole feature is vestigial and Bethesda basically doesn't even use it, lol). Fixed padding on the control for editing a skill offset, too.

The ActorBase UI now does all stat calculations, stores the results on the form, and then synchronizes them back into the UI.

The ActorBase UI should now react to the actor's race or class having changes committed from afar, and recalc stats.

The backend stat calculation now makes it easier to see the base values of each stat, and not just the computed ones. Only the UI needs this.

Added typedefs to the backend for the value type of an ActorBase's skills and attributes. Made attributes a bit easier to work with.

NOTE: We currently don't auto-calc skills properly. Not sure what's going wrong yet.