DKFloatSlider, DKTabWidget, DKFormPicker fix, ActorBase UI
Finished DKFloatSlider and integrated it into Qt Designer. The ActorBase UI now uses it for the tint layer alpha slider and for the morph sliders.

Implemented DKTabBar and DKTabWidget, and made the latter available in Qt Designer, to address a bug with QTabWidget: if you set a tab as hidden, you can still navigate to that tab by using the scroll wheel on the tabbar.

DKFormPicker: Fixed widget sizing further; the widget should no longer collapse to an unusably small size when placed in a layout with a spacer item (and when nothing else is present to "protect" the picker's size). The widget will still allow itself to collapse and truncate its items when the layout actually requires it.

ActorBase UI: added a face morph slider for VampireMorph.

ActorBase UI: we now update the face morph sliders when template flags change.

ActorBase UI: the Add and Remove buttons for the faction relationship list now work.

ActorBaseFactionsModel: we now properly update the cached editor ID for a faction when overwriting a faction relationship.