Warning log and load window UI/UX improvements
The main window's status bar now has a permanent widget which shows the number of currently logged warnings. When new warnings are logged, this widget flashes orange a few times. It also indicates the number of unread warnings, with this count updated when you focus the Log Window.

Moved the model for the Log Window out of the UI side of things and into the message-log subsystem.

When loading game files, the title of the window you use now indicates which game you're loading files for.

Started (but only barely) work on editing tint layers in an ActorBase.