ActorBase work; classed stat computation fixes; DKFormPicker fix
ActorBase UI now supports editing tint layers.

ActorBase UI now prevents you from setting an actor as Unique if they derive (via template actors) from a leveled actor. It'll also warn you (and clear the flag) when opened, if they already do and are flagged as Unique.

Moved some enums outside of the loaded forms that defined them, and into their own files.

Fix: if an actor's class defines no weights, we no longer crash on an integer division by zero when computing their classed stats; like the game, we just don't assign any points.

ActorBase UI: We no longer try to maintain separate data for male and female actors. It would've been nice to make it so you don't lose any data if you switch an actor's sex back and forth, but unfortunately Qt made this a pain in the ass: changing a view's model also nukes the selection model (and any signal/slot connections we made to it), and I couldn't get any workarounds to this working (because Qt is effectively opaque to my debugger, so I can't tell where things are going wrong when trying to finesse my way around its internals).

DKFormPickerModel: fixed a lifetime bug. We don't own the custom filter (because it could be applied to multiple pickers) so we have to account for the possibility of it being destroyed before we are.