ActorBase UI work: HeadParts
HeadParts are now listed in the ActorBase dialog. Base head parts are editable; we still need to add drag-and-drop and support for the Delete key for extra head parts. HeadParts properly update as the actor's race and sex changes and as the template actor changes, and should be saved back to the form properly. (We will miss changes to a HeadPart's race FormList, but c'est la vie.)

We now handle the `NPC_/NAM5` subrecord, and so should no longer emit spurious warnings about it.