ActorBase UI work
Improved how FaceGen data is stored on ActorBase forms in the backend. It's grouped together rather than split in two now, and the morph indices and morph sliders are more clearly named and can be indexed like arrays.

We should now only save `NPC_/NAM9` if any of the sliders is non-zero.

Fixed bugs in how we sync data from the UI to a working-copy ActorBase's face tint layers.

The Base Head Parts and Additional Head Parts lists in the ActorBase dialog now support drag-and-drop. The latter list also allows deletions via the Delete key, and applies the same constraints to drag-and-drop as in the CK.