Work on the Race UI
Added buttons to add, reorder, and delete presets within a tint layer. Refactored how we handle sex-differentiated UI for editing face tints. Refactored how we push changes from the UI to the tint models.

The dialog now forces all tabboxes' current tabs on startup. (Qt Designer uses the last-viewed tabbox in the editor as the default tab in your actual program, which is... not great.)