DKFormListPane custom filters; Race preset filters
DKFormListPane now supports custom filters. These are very similar to those available for DKFormPicker, except that the pane custom filters are allowed to fully load the form as part of filtering. To be perfectly blunt: this was done out of laziness to avoid having to add ActorBases to the form info cache; but it's also fine because list panes should never have as many forms as a formpicker may potentially have.

The Race UI's lists of preset actors now use custom filters, to only allow actors of matching race and sex.