DKFormListPane enhancements; backend fixes; Race M.D.O. editing
DKFormListPane now emits signals when its selection changes, and allows retrieval of its selection. It also allows you to configure whether multi-selection is enabled.

Fixed mistakes in the RACE backend: we weren't properly loading or building use info for MTYP+SPED.

As the movement type override data in Race is of the same basic format as the base data in MOVT, I've moved it to an external struct shared by both forms.

Fixed a backend mistake: I'd forgotten to add MovementType to the all-forms class list, so we weren't able to load it properly.

Implemented support for editing Races' movement data overrides.