Quest dialogue UI work
Refactored QuestAllDialogueModel and excluded it from the project. We now use a backing "datastore," and four separate Qt models to adapt it to the view. Note that word: "adapt." The primary purpose of a Qt model is to adapt one type of data for Qt's various view widgets, though a Qt model can also store its own data. In this case, we have three different types of data; a single Qt model can't handle that in a maintainable, scalable, clean way; things get out of control too quickly, as they did with the now-unused QuestAllDialogueModel.