Fixes for deleting forms; work on dialogue editing UI
Fixed a Dovahscript crash when deleting a form when Dovahscript wasn't actually running. (There was an assertion to safeguard against forms being deleted out from under a running script, but I, uh, forgot to actually check whether a script was running.)

Fixed a crash when deleting a form while a form-editing dialog was open for it.

Added a form-editing dialog for Dialogue Branch forms.

Opening a form-editing dialog for dialogue branch, topic, or info forms should now: open the quest-editing dialog, or focus it if it's already open; navigate the quest-editing dialog to the appropriate tab views, and select the appropriate dialogue forms in the listviews; and then open the new form-editing dialog for the dialogue form you wish to edit.

The dialog box for deleting forms should now have a correct window title, and the list of using forms should no longer include any of the to-be-deleted forms (including their child forms).