Prep for supporting co-opting of the hardcoded form ID range
Made some backend changes to prep for supporting cannibalism of the hardcoded form ID range (the new behavior for record IDs introduced in an SSE update). Unfortunately, full support is not possible at this time: we need to be able to map every file's local IDs based on its master list in order to load form data properly, but we also need to be able to map the active file's form ID range based on what's permitted in memory (i.e. the masters it *will* have when we save) in order to determine what IDs are legal with respect to form creation and numbering. Currently, *all* form ID mapping relies on `file_prefix` objects that are constructed in the same way for files: there's no distinction between "masters" and "runtime."

TODO: prefer "co-opting" over "cannibalism."