Refactored the form-info-cache subsystem
Refactored the form-info-cache subsystem. The types are now better-organized, and the subsystem now uses the PImpl idiom, which should hopefully avoid unnecessary recompiling when I make changes to it.

The FIC should now maintain a list of all Topics that are in the IDAT (SharedInfo) subtype. We can use this to find all SharedInfos across all quests without having to fully load any forms.