Numerous bugfixes for editing dialogue
Fixed a bug with DKCustomFormFilters that could cause them to fail: they static-casted when they needed to dynamic-cast to access a multiply-inherited parent class.

DKCustomFormFilters now prune their internal list of tracked models, so that if a model is deleted from afar, the list doesn't just grow progressively longer and longer with null entries. (This is mainly a concern for if you're using filters on disposable DKFormPickerDialogs.)

The models for DKFormPicker and DKFormPickerDialog now properly call the QAbstractItemModel superclass, which is needed for QObject parent/child relationships and memory management to work properly.

The model for DKFormPickerDialog now actually uses its custom filter, if one is present.

QuestDialogueTabBody no longer allows you to delete the starting topic of a branch.

The UI for editing a Topic no longer allows you to change the topic's subtype from SharedInfo if it contains any INFOs that are currently being used as SharedInfos by other INFOs.

These fixes resolve memory leaks as well as bugs with adding linked topics to an INFO (the program would hang because it'd try to generate a dialog with every single loaded DIAL inside) and with setting an INFO's Walk Away Topic (the formpicker would never fill, because it wasn't invoking the custom filter properly in response to model changes).