Partial fix to handling ordered INFOs within a DIAL
Our existing code wasn't designed to properly save reordered INFOs within an overridden DIAL (i.e. the active file reordering INFOs within a DIAL defined in one of its master files). I've made code changes to attempt to remedy this, though they're not yet tested. A DIAL's form stub addenda now maintains two lists of ordered children, kept synchronized during load: one is the list of ordered children as defined in the active file's masters, while the other is the list of ordered children as defined in the active file itself. We modify the latter list while editing, and when we save a file, we diff the two lists to know whether and how to write INFO/PNAM.

There is a defect in the code inside of this commit: while loading the active file itself, we would still keep the two lists synchronized even when we should only be modifying the active file ordered child list at that point. I'd wait to fix it before committing, but I just had a power outage that nearly bricked my PC, and I'd rather not lose today's work.

Fixed an issue with the file writer: it had no way to explicitly write a zero INFO/PNAM ID, because it relied on a form stub pointer to know whether to write PNAM at all. We now use std::optional<form_stub*> so we can distinguish between "no PNAM" and "empty PNAM."

Took this opportunity to clean up a few other things, like grid positions in cell addenda.