Attempted fix for previous commit
Attempted to fix the issue mentioned in the previous commit's description, regarding managing a DIAL's ordered-child list when loading the active file.

Modified the threaded loader for basic forms to remove some dead code: upon encountering an INFO, it will not check the current group to see if said group is a topic info group. It can't be: the basic-form threaded loader is explicitly designed not to handle topic info groups (the dialogue threaded loader does this instead).