UI designs for form types
UI designs (no backend or frontend implementation yet) for the following form types: CAMS, COBJ, LSCR, MATO, SPGD, SOUN.

For the COBJ UI, we'll need to make DKFormInventoryPane capable of hiding ExtraData-related widgets (health; ownership), make it capable of disallowing leveled items, and make it hide "Preview Calculated Result" widgets when leveled items are disallowed. BGSConstructibleObject uses the TESContainer component -- the inventory component -- so we need to represent the crafting recipe as an inventory in the backend, but the UI should be capable of displaying just those options that can be used sanely.