Major stability fixes
Fixed a major issue in the design of `cobb::multiheap`: there was nothing preventing the case of one thread attempting to free an object while another thread allocated one, creating a risk of memory corruption. DovahKit worked 99.999% of the time because form stubs were only allocated or freed on multiple threads during file load, and only freed in the case of loading an override. The one (1) incidence of memory corruption linked to the multiheap occurred a week or few ago (I was unable to debug it due to a breathtakingly stupid Visual Studio UI bug that cost me all relevant debugging information). Hopefully this commit will fix whatever caused it.

Separately, handling for the ARMO/ARMA override warning has been fixed. We can't store the overriding form stub because it gets deleted. (Form stubs are heap-allocated by the threaded loaders and then passed to the `file_load_order` object. The FLO keeps a form stub if it's new, or, for overrides, merges the new stub's data into the existing stuf and deletes the new stub.)

Updated the form stub Natvis to match the recent changes to form stub addenda.

Fixed a bug wherein we mishandled the ordered child lists of DIALs when an INFO was reparented from one to another.