DKQuestSceneEditor cleanup
Split the various components of DKQuestSceneEditor into their own files, so the code is more manageable.

Changed the components' geometry to use QRect rather than a paired QPoint and QSize.

Moved the code for calculating an actor box's size out of DKQuestSceneEditor itself and into the DKQuestSceneEditor_impl::Actor class.