Scene editor: dragging actions between actors/phases
It is now possible to click and drag on an action in the scene editor, to drag it to another phase or even another actor, so long as the target actor-and-phase isn't already occupied by another action of the same type.

Fixed a bug that could happen when switching the quest UI back and forth between the same scene without making any changes. We were basically leaking the scene's working-copy form data. (God, I can't wait until I do away with that whole system as part of sustain...)