Scene editor: mouse panning, cursor updates
It's now possible to pan the scene editor by clicking and dragging on anything that wouldn't otherwise be interactive.

The mouse cursor now updates depending on what part of the scene editor it's placed over, and whether you're drag-panning. This includes showing resize cursors on scene actions, though drag-resizing is not yet implemented.