Scene editor: drag-to-resize actions
It's now possible to click and drag the edges of an action to resize it across multiple phases. The implementation is still unfinished: the way we display the action while it's being resized is imperfect (we can display it at sizes smaller than it's allowed to become when you mouseup, including negative sizes).