Scene editor: zooming
The zoom buttons are now functional. The internal geometry of the things in the editor isn't zoomed. Rather, the editor applies zooming at paint time, and for mouse hit detection, we just "un-apply" the zoom to the widget-relative or global mouse position, to map them to the un-zoomed internal geometry.