DKYesNoUnsetWidget; Qt Designer fix-up
Created the DKYesNoUnsetWidget, intended for Package procedure flag overrides.

Added icons to some of the custom widgets in our Qt Designer plug-in, to make them easier to spot in the widget listing.

Fix for compiling the Qt Designer plug-in: Qt changed the macro we need out from under us, so now we're using a more appropriate one.