DKGameFilePicker refactor
The UI's handling of game file paths is now (in at least some places) abstracted behind ui::types::game_file_path, a class similar to std::filesystem::path but designed to let us convert to the specific representations we need for the TES engine.

DKGameFilePicker has been redesigned to store paths as ui::types::game_file_path, and simply alter how they're displayed. If outside code wants paths in a different format, they have to request that directly. This is an improvement over the prior status quo, wherein DKGameFilePicker's path conventions were poorly documented and generally ill-conceived.

Updated all UIs that use DKGameFilePicker, to match the new interface. Updated the Qt Designer plug-in as well.

Overall, the ergonomics for working with these file paths are poor, largely because it's something I'm grafting on so late into the project. This is something we can deal with during sustain. File paths should probably be a backend type (and should probably be better-designed; among other things, we can't distinguish between the absolute path "Data/" and the relative path "Data/").