more
See previous two commits.

Changes to DKBSABrowseDialog were meant to make it into the first of those, but since I forgot, I'll describe them here. In addition to adjusting the thing to match some changes to DKGameFilePicker, these changes should also fix it not opening to the right folder (the path stem) when not given an initially selected file.