MGEF mostly done; fixes for IPCT, REVB, VMAD
MagicEffect forms are now mostly implemented. There are still some bugs that need fixing:

* We load the description improperly (may be a `common` lstring rather than `description`, since it's DNAM rather than DESC).
* The UI doesn't properly display the Resist Value for any MGEF; it's always "None" upon opening.
* The UI should use "NONE" rather than "None" for AV dropdowns, for consistency with DKFormPicker. Potentially we may want to make a custom widget for AV dropdowns anyway.
* The "Snap to Navmesh" checkbox needs to be greyed out when the delivery type is not "Target Location."
* MGEF UI: When an associated item's type is "actor value," we should show the actor value names, not the AVIF editor IDs. This may require changes to DKFormPicker (maybe a pref that can be set on it, defaulted to false, to prefer AV names over AV editor IDs).

---

IPCT fix: we now properly clone the "impact result" value.

REVB: Reflect Delay is now scaled within the backend rather than the frontend, and consistently with the CK.

Critical fix to the VMAD loader: if the subrecord has only one script with no properties on it, we no longer bounds-check it improperly and fail to load the script.