Form-info-cache subsystem refactor
The form-info-cache subsystem previously relied on "cacheable trait" structs which defined metadata associated with cacheable traits, such as the form types of interest. It was designed this way because not all cached data has a unique type; for example, some cached data is just bare QStrings. However, it meant that when cacheable data *did* have its own type, information related to what we cache and when was dispersed across different files in different folders. The "cached data" structs per form type have been extended to now supply the information that used to be tied to "cacheable trait" structs, to remedy this.

Additionally, we now use template metaprogramming to cut down on some code repetition, making use of cobb::class_array and template lambdas.