ENCH UI; SPEL UI fixes
Implemented a UI for ENCH forms.

Fixed oversights in the SPEL UI regarding the Auto-Calc flag.

Missing change from a previous commit: frontend text for IDLE load warnings.

Added descriptions for the fEnchantingRuleOf... GMSTs.

The backend's "magic effect list item cost calculator" object now caches the loaded Magic Effect's charge time. This isn't used for computing the cost of a single Magic Effect, but is useful when performing Auto-Calc for the charge time of an entire effect list, which is a closely related operation.

Reorganized the backend loader for SPEL, to have formatting and code comments consistent with ENCH and ALCH.