SOUN legacy data; EFSH backend fixes; UI fixes
We now load legacy data for TESSound forms.

EffectShader forms now use the correct typing for animation frame counts.

If you take a DKFormPicker that only accepts Sound Descriptor forms, and bind it to a form_reference_t that currently refers to a Sound form, then when updating the formpicker's initial value, we now follow the Sound to the wrapped Sound Descriptor.

The EffectShader UI will now display textures without their alpha channels if the relevant flags for disabling alpha are set.

Fixed bugs with binding a DKGameFilePicker to a target path, and to a DKTextureAssetPane.