CSTY, MATO, MATT
Fixes to EFSH.

Implemented CSTY, MATO, and MATT. The implementation of MATO is partial (see comments in the class definition); we can't fix that absent a rebuild of our NIF code (which I guess will be yet another thing to handle in Sustain Phase 1), which we need for other (sustain) reasons anyway. We can load and save MATO data, but editing is limited.

STAT should now auto-update visibility of SSE-only options if you convert between games while the dialog is open.

Drafted up plains for a Sustain Phase 1 redesign of our NIF-handling code.