PACK backend work
The backend for TESCustomPackageData is still incomplete, but I have the skeleton written up -- all the individual parts are here.

In Skyrim's engine, Idle data in a PACK and the data of an IDLM are a shared form component: BGSIdleCollection. Split this off into a reusable component in DovahKit as well.

Split package locations and package targets into a separate type; treating them as uniform was a mistake. Did more RE and identified more fields on package locations.