PACK backend work
Serialized package data is split in half. Previous drafts of our PACK backend unified these, but I realized why Bethesda themselves split them up: it's values and then declarations, and if a package uses a template, then it has no declarations of its own and instead uses the declarations on the template. I've recreated that split, both to make it simpler to implement support (i.e. less to do all at once) and to make it easier for code outside the backend to eventually work with package data (i.e. we can have functions that take two value maps and a declaration map, so when a package uses a template, you can pull each from the approriate places).