PACK backend work
Custom package substructs now use a subordinate UIB instead of a vector of form IDs. This change was necessary because procedure tree nodes can have conditions, which can only generate use info into a UIB.

Created a dedicated struct for package procedure trees. This struct is responsible for converting between a flat list of procedure tree nodes (as is present in a data file) and an in-memory tree.

Noted some design flaws in the use info generation system (well, not the system as such, but the structs that we use to do the thing) and wrote up Sustain Phase 1 plans to fix them.