PACK backend fixes
These are based on tests against [PACK:19713] in Skyrim.esm. With the changes in this commit, we now load that form correctly.

Corrected the enum width for PLDT and PTDA's types.

Corrected the control flow for opening subrecords versus reading the current subrecord.

Corrected unflattening of the procedure tree, w.r.t. detecting orphans.