NAVI backend now ready to build and test
NavMeshInfoMap now stores road markers using a container that acts like a map, instead of in a vector that allows duplicates. I say "a container that acts like a map" because form_reference_t cannot be the key type of an STL map container (or at least, if you use it that way, you'll very quickly run into some nasty problems writing the boilerplate for the enclosing form class, since you don't have full control over how the keys are created or destroyed).

All the parts of NAVI are implemented; now I just need to compile and test it.