Backend: LENS, VOLI
Added backend support for the form types added  by Skyrim Special Edition.

At this point, the only form type left to finish is LCTN (Location), which we have incomplete code for. We need to review the internals more thoroughly, to figure out how to distinguish the various bifurcated lists therein.