Backend: LCTN
Implemented the parts of LCTN that weren't previousy supported within the backend, i.e. the six content lists.

This was, frankly, terrifying, and I'm not fully confident that I have it right. It's very challenging to try and map how the CK and game handle this data, to how DovahKit has to handle this data.