Perk work
Implemented the UI for editing Perks and their entry points. Made some minor fixes to the backend, and improved how we represent entry points within DovahKit (though further improvements are needed there).

Made some UB fixes to some of the self-sorting UI models. In the next commit, I plan to make more use of a helper function that I already had and am using in the model for perks' entries.