Object Window categories
Actor Values, and the SSE-only Lens Flare and Volumetric Lighting form types, are now listed in the Object Window. This is a bare-minimum edit; in particular, neither the UI nor the backend currently take any steps to prevent the user from creating forms of these types when it would be inappropriate to do so (e.g. creating SSE-only forms in an LE file; creating Actor Values ever).

(The Skyrim Creation Kit puts all Actor Values in one shared dialog, with a listview on the side to select an AV to edit. By contrast, we now list AVs in the Object Window consistently with Fallout 4 so that each AV can get its own dialog. As much as possible, I want the UI behavior to be that if you click "Cancel" when editing a form, then any changes you made in the dialog are undone. This is significantly more complicated to do when implementing a single dialog that involves switching between editing multiple forms; it's a major pain point when working with Scenes, Topics, and Topic Infos, which are all accessed from within a Quest dialog, for example. Putting AVs in the Object Window means you can (and must) open an individual dialog for any AV you want to edit, and so that dialog can have a trustworthy "Cancel" button.)